requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-extensions==3.2.3
idna==3.10
MarkupSafe==3.0.2
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
sslcommerz-lib==1.0
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.1.3
